Fix test/fsa

+++ test/fsa/src/adfa/nest_louds_trie_test.cpp
@@ -52,7 +52,7 @@ GetoptDone:
 	}
 	NestLoudsTrie_SE trie;
 	SortableStrVec strVec2 = strVec1;
-	valvec<uint32_t> linkVec(strVec2.size(), UINT32_MAX);
+	valvec<size_t> linkVec(strVec2.size(), UINT32_MAX);
 	trie.build_strpool(strVec2, linkVec, conf);
 	if (b_write_dot_file)
 		trie.write_dot_file((std::string(fname) + ".dot").c_str());
